pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "learningPyTorch"
version = "0.1.0"
dependencies = [
    "torch",
    "torchvision", 
    "tqdm",
    "numpy",
    "scikit-learn",
    "requests",
    "matplotlib",
    "torchinfo",
    "Pillow",
    "ipykernel",
    "ipywidgets",
    "tensorboard",
    "torch-tb-profiler",
    "pandas",
    "pytz",
    "tzdata",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["modules*"]